Several fixes. See details below. ----------------------------------------------------------------------- TESTING - Straightforward. ----------------------------------------------------------------------- DETAILS - fAbout.Designer.cs - fAbout.resx   * Updated version to 2.3.0.1   * Updated copyright.
- fGridFilter.cs
- fReport.cs
  * Fixed another Toolbox.DBI bug (long/int casting)

- fMain.Designer.cs
  * Removed stray separator from tools menu.
  * Fixed "Edit.." bug.

- fMain.cs
  * Changed Report, Grid, Punch Card dialogs to be non-modal.
  * Disabled the File menu when the timer is running.
  * Re-enabled it when the timer stops.  :)
-----------------------------------------------------------------------
Project...: Timekeeper
Components: n/a
Target....: 2.3
Ticket(s).: 1173, 1179, 1215

[git-p4: depot-paths = "//depot/timekeeper/": change = 215]

+++ timekeeper/fGridFilter.cs
@@ -43,7 +43,7 @@ namespace Timekeeper
             foreach (Row row in rows)
             {
                 Pair item = new Pair();
-                item.Key = row["id"];
+                item.Key = (int)row["id"];
                 item.Value = row["name"];
                 list.Items.Add(item);
             }
